requirements.txt
Django>=4.2.13,<5
resolver==0.2.1
selector==0.10.1
Paste==3.10.0
PasteDeploy==3.1.0
PasteScript==3.6.0
feedparser==6.0.8
Markdown==3.6
smartypants==2.0.1
psycopg==3.1.19
olefile==0.47
versiontools==1.9.1
statsd==4.0.1
pep8==1.7.1
pyflakes==3.2.0
mccabe==0.7.0
pycodestyle==2.12.0
configparser==7.0.0
importlib-metadata<7.2 # for flake8
typing_extensions==4.12.0
zipp==3.19.0  # for flake8
flake8==7.1.0
tornado==6.4.1
cssselect==1.2.0
lxml==5.2.0
fuzzywuzzy==0.18.0
sure==2.0.0
extras==1.0.0
python-mimeparse==1.6.0
fixtures==4.1.0
testtools==2.7.1
testscenarios==0.5.0
python-subunit==1.4.0
coverage==7.5.0
pyasn1==0.6.0
pyOpenSSL==24.1.0
ndg-httpsclient==0.5.1

chardet==5.2.0
idna==3.7
urllib3==2.2.2  # sentry-sdk
certifi==2024.6.2  # sentry-sdk

requests==2.32.1
logilab-common==2.0.0
astroid==3.2.0
editdistance==0.8.1
pylint==3.2.0
sqlparse==0.5.0
s3transfer==0.10.0
jmespath==1.0.0

six==1.16.0  # boto3

boto3>=1.12.0,<1.35.0
botocore>=1.15.0,<1.35.0

python-dateutil==2.9.0
contextlib2==21.6.0
Jinja2==3.1.4
MarkupSafe==2.1.1
pytz==2024.1

typogrify==2.0.7
django-staticmedia==0.2.2
django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-appconf==1.0.5
django-statsd-mozilla==0.4.0
raven==6.10.0
sentry-sdk==2.6.0
django-bootstrap-form==3.4
django-debug-toolbar==4.3.0

django-smoketest==1.2.0
django-extensions==3.2.3
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-storages==1.14.2
gunicorn==22.0.0
django-impersonate==1.9.1
django-smtp-ssl==1.0
cffi==1.16.0 # cryptography
cryptography==42.0.4  # pyOpenSSL

ctlsettings==0.3.1

pbr==6.0.0
PyYAML==6.0.1 # MIT
stevedore>=1.20.0 # Apache-2.0
bandit==1.7.1

entrypoints==0.4
asgiref==3.8.0

backports.zoneinfo;python_version<"3.9"
webencodings==0.5.1 # django-markdownify
tinycss2==1.3.0 # django-markdownify
bleach==6.1.0 # django-markdownify
django-markdownify==0.9.5
